requirements.txt
django==4.1.7
djangorestframework==3.14.0
django-environ==0.10.0